Move ARI model from recurrent/parametric to recurrent/renewal
The ARI (Arithmetic Reduction of Intensity) model is a renewal/repair model
alongside ARA and the generalized renewal models, so it belongs in the
renewal folder rather than the parametric (baseline intensity) folder.
