+++ surpyval/recurrent/__init__.py
@@ -1,8 +1,9 @@
 from .nonparametric import NonParametricCounting
-from .parametric import ARI, HPP, CoxLewis, CrowAMSAA, Duane
+from .parametric import HPP, CoxLewis, CrowAMSAA, Duane
 from .regression import ProportionalIntensityHPP, ProportionalIntensityNHPP
 from .renewal import (
     ARA,
+    ARI,
     GeneralizedOneRenewal,
     GeneralizedRenewal,
     RenewalModel,
