+++ surpyval/recurrent/parametric/__init__.py
@@ -3,6 +3,3 @@ from .crow_amsaa import CrowAMSAA
 from .duane import Duane
 from .hpp import HPP
 from .parametric_recurrence import ParametricRecurrenceModel
-
-# ARI imported last: it builds on the baseline intensity models above.
-from .ari import ARI  # noqa: E402
