+++ surpyval/recurrent/renewal/__init__.py
@@ -2,3 +2,7 @@ from .ara import ARA
 from .generalized_one_renewal import GeneralizedOneRenewal
 from .generalized_renewal import GeneralizedRenewal
 from .renewal_model import RenewalModel
+
+# ARI imported last: it builds on the baseline intensity models in
+# ``recurrent.parametric``.
+from .ari import ARI  # noqa: E402
